Add sending of FullStateAck event to GC
Sent in response to every FullState message received. Help GC know the
connection is alive. Used (at least initially) to handle when GC headset
pauses or is out of bounds for long enough that RCapp timesout (happens
automatically somewhere in Peer I believe), and when GC comes back
to action, it has to recognize this and reset.